refactor(bot): extract bot/agent_loop.py — closes #121
Final slice of the bot/core decomposition (PRD #113, ADR 0001). Carves
out the multi-round LLM dispatch loop + its supporting helpers from
``bot/core.py`` into ``bot/agent_loop.py``, completing the refactor
that began in PR #122. **``bot/core.py`` is now 661 LOC (−87.2% from
the 5160-LOC baseline on main pre-PR #122).**

Module surface (18 re-exports):

* ``llm_tool_loop`` — the central multi-round dispatch coroutine that
  every User-facing entry walks through. Signature unchanged from
  pre-extraction (verified by the new
  ``test_llm_tool_loop_signature_matches_production_call_sites`` test).
* ``SYSTEM_PROMPT`` module global + ``_ensure_system_prompt``.
* Tool registry hooks: ``get_tools``, ``_build_bot_tool_context``,
  ``get_tool_registry``, ``_build_llm_timeout``.
* ``MAX_TOOL_ITERATIONS`` configurable cap.
* Per-call adapters: ``_format_llm_api_error_message``,
  ``_sanitize_tool_history``, ``_normalize_tool_callback_args``,
  ``_emit_tool_callback``, ``_coerce_timeout_seconds``,
  ``_extract_timeout_seconds_from_text``,
  ``_extract_tool_timeout_seconds``,
  ``_build_tool_result_callback_metadata``,
  ``_build_bot_query_engine_callbacks``,
  ``_maybe_resume_pending_preflight_request``.

Cross-module access:

* Stable bot.core symbols (``OUTPUT_DIR``, ``OMICSCLAW_DIR``,
  ``transcript_store``, ``tool_result_store``,
  ``pending_preflight_requests``, ``audit``, ``MAX_HISTORY``,
  ``MAX_CONVERSATIONS``, ...) imported at module top.
* Runtime-reassigned globals (``llm``, ``OMICSCLAW_MODEL``,
  ``LLM_PROVIDER_NAME``, ``memory_store``, ``session_manager``)
  accessed via ``_core.<name>`` at call time so they see the values
  ``bot.session.init()`` writes.

**Lazy re-export pattern in bot.core**:

To avoid a circular load when ``bot.agent_loop`` is imported before
``bot.core`` (test path or downstream module), the agent-loop names
are NOT eagerly re-exported in ``bot.core``. Instead,
``bot.core.__getattr__`` lazily resolves them on first access via
``import bot.agent_loop``. The 18-name set is enumerated in
``_AGENT_LOOP_REEXPORTS`` at the top of the ``__getattr__`` block.
External code reading ``bot.core.llm_tool_loop`` or
``bot.core.SYSTEM_PROMPT`` continues to work unchanged because Python
falls through to ``__getattr__`` when the name isn't in module
globals.

Tests added (``tests/bot/test_agent_loop.py``, 6 cases):

* Tracer bullet: ``llm_tool_loop`` lives at canonical ``bot.agent_loop``
  and is a coroutine function.
* Backward-compat: ``bot.core.llm_tool_loop is bot.agent_loop.llm_tool_loop``
  (lazy re-export resolves to same object).
* Production-callers contract: ``llm_tool_loop`` accepts every kwarg
  that ``omicsclaw/app/server.py`` and
  ``omicsclaw/interactive/interactive.py`` pass.
* Pure-helper behaviour: ``_coerce_timeout_seconds`` (int / str /
  float / negative / None handling), ``_extract_timeout_seconds_from_text``
  (timeout-phrase recognition), ``_format_llm_api_error_message``
  (non-empty actionable text for arbitrary exceptions).

**Deferred** (per #121 acceptance, but not in this commit):

The full integration tests against a fake ``AsyncOpenAI`` client
(single-round / multi-round tool dispatch / inspection-only round /
streaming callback ordering / API-error path) require substantial
fixture work — the loop takes 25+ kwargs and orchestrates many
collaborators. The 3 pure-helper tests above plus the existing
behavioural-parity eval suite (T9, ``tests/eval/`` against
``deepseek-v4-pro``) cover the contract for now. Heavy integration
tests are tracked as a future follow-up; this commit ships the
extraction safely on top of the existing regression net.

Verification:

* New ``tests/bot/test_agent_loop.py``: 6 / 6 pass.
* Bot regression net: 234 passed (+6 new), 1 skipped, 4 xfailed —
  no regressions across ``test_bot_*``, ``test_oauth_regressions``,
  ``test_app_server``, ``test_notebook_files``,
  ``test_routing_regression``, ``test_auto_routing_disambiguation``,
  ``test_skill_listing``, ``test_bot_core_timeout``, ``tests/bot/``,
  ``tests/runtime/preflight/``.
* ``tests/eval/`` + ``tests/test_knowhow.py``: 121 passed, 17
  deselected. Behavioural-parity baseline 14/15 against
  ``deepseek-v4-pro`` unaffected.
* ``bot/core.py``: 1441 → 661 LOC (−780). **Cumulative drop from main
  baseline: 5160 → 661 LOC (−4499, −87.2%).**
* ``bot/agent_loop.py`` lands at 908 LOC.

Closes #121.
